refactor: extract RPCHandler to a separate file
